Switch from float to double for variables and use of stream instead of for iteration

+++ excel-process/src/main/java/model/Person.java
@@ -9,9 +9,9 @@ public class Person extends AbstractPerson {
 
     @Getter
     @Setter
-    private float payRate;
+    private double payRate;
 
-    public Person(String name, String surname, Date birthDate, float payRate) {
+    public Person(String name, String surname, Date birthDate, double payRate) {
         super(name, surname, birthDate);
         this.payRate = payRate;
     }
